kvm: Fix for dead lock while rebooting VM
Cherry pick for tests purposses. It should be merged in
pull request #2191

This patch is fixing dead lock issue in lkvm. When kernel is sends
keyboard signal for reboot to lkvm, it sends SIGKVMEXIT to all cpu threads.
When first cpu became closed, devices are checking which cpus are still
alive and send to them SIGKVMPAUSE signal(for detaching, devices require
threads in pause or exited state). After this signal, devices are waiting
for response from threads.

Main problem in this bug is situation when cpu thread becames exited and
SIGKVMPAUSE signal was sent. In this situation Pci-device is still waiting
for notification from closed thread.

To fix this bug I am waiting for exit status from all cpus before, turning
down devices.